requirements.txt
python-telegram-bot==20.3
PyDrive==1.3.1
PyYAML==6.0
google-api-python-client==2.80.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0